refactoring: use size_t when it definitelly known, that a number of bytes assumed

+++ heap.h
@@ -1,6 +1,8 @@
 #ifndef __HEAP_H__
 #define __HEAP_H__
 
-void* malloc(unsigned size);
+#include <stddef.h>
+
+void* malloc(size_t size);
 
 #endif
